Corrigir troca de app fora do loop, parser de OPF e duplicação no OTA

+++ lib/Book32_Core/AppMgr.cpp
@@ -12,7 +12,7 @@ void AppMgr::registerApp(App* app) {
 }
 
 void AppMgr::switchTo(int index) {
-    if (index >= 0 && index < apps.size()) {
+    if (index >= 0 && index < (int)apps.size()) {
         if (currentApp) currentApp->stop();
         currentApp = apps[index];
         currentApp->start(); // Use start() instead of setup()
